group1/pyproject.toml
[project]
name = "group1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "langserve>=0.1.1",
    "faiss-cpu>=1.10.0",
    "langchain>=1.1.3",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "numpy>=1.26.4",
    "openai>=1.109.1",
    "pandas>=2.3.2 ",
    "pydantic>=2.11.9",
    "python-dotenv>=1.2.1",
    "fastmcp>=2.13.3",
    "langchain-community>=0.4.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "mcp>=1.22.0",
    "langchain-mcp-adapters>=0.2.1",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "jupyter>=1.1.1",
    "pyright>=1.1.407",
]

[tool.autopep8]
max_line_length = 88
ignore = "E203"

[tool.pytest.ini_options]
pythonpath = "."
norecursedirs = "tests/_helpers"
testpaths = "src tests"
env = ["IS_TEST=true", "KEY_VAULT_NAME=testvault"]

[tool.pyright]
typeCheckingMode = "basic"
